+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added FEP-ae97 API endpoint for deleting media.
 - Added FEP-ae97 gateway metadata endpoint.
 - Set `discoverable` flag to `true` on actors.
+- Specify minimum payment amount in `Proposal` objects.
 
 ### Changed